Rename API endpoint for validators to validation
