+++ st2api/st2api/controllers/exp/root.py
@@ -15,7 +15,7 @@
 
 from st2api.controllers.exp.actionalias import ActionAliasController
 from st2api.controllers.exp.aliasexecution import ActionAliasExecutionController
-from st2api.controllers.exp.validators import ValidationController
+from st2api.controllers.exp.validation import ValidationController
 
 
 class RootController(object):
@@ -26,4 +26,4 @@ class RootController(object):
     aliasexecution = ActionAliasExecutionController()
 
     # Experimental
-    validators = ValidationController()
+    validation = ValidationController()
